VehicleWheelQueryResults in separate class
